LPS-72687 Moves tests
